Created two new FaceSelections to fix errors caused by FacingCenter fix
FaceSelections now has two new entries:
- FacingUpCenter
- FacingDownCenter
These entries are to both fix the gyroelongated cupola and bicupola as it was using the FacingCenter bug to work, and also to help work with antiprisms easier. The FaceSelections use a conditional logical AND operator to compare if it faces both up and to the center of the shape.